apps/workers/diarization/requirements.txt
pyannote.audio==3.1.1
torch>=2.1.0
torchaudio>=2.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
